fix(build): bump cmake_minimum_required to 3.5 for CMake 4.1 compat
CMake 4.1.3 (now in vitasdk/vitasdk:latest) dropped backwards-
compatibility support for cmake_minimum_required VERSION < 3.5.
The previous floor of 3.2 caused a hard error at configure time.
3.5 is the new minimum; no policy or feature changes are needed.

Also fix three GCC 15 hard errors exposed once CMake configure succeeded:
- lib/src/discovery.c:178: add explicit cast (struct sockaddr_in6 *)
  to match the pattern used on lines 171 and 189 of the same file.
- lib/src/ffmpegdecoder.c: remove avcodec_close() call removed in
  FFmpeg 6+; avcodec_free_context() already closes the context.
- vita/src/video.c: add extern declarations for six codec-engine /
  extended-videodec functions whose symbols exist in VitaSDK stub
  libs but have no header declaration (GCC 15 errors on implicit
  declarations).
- vita/src/message_log.c: include <psp2/kernel/processmgr.h> to
  supply the declaration of sceKernelGetProcessTimeWide().
- vita/include/message_log.h: widen start_offset/lines from uint8_t
  to uint16_t (MLOG_LINES=500 overflows uint8_t; GCC 15 warns).

+++ vita/include/message_log.h
@@ -10,8 +10,8 @@
 // A hard-wrapped message log
 typedef struct vita_chiaki_message_log_t {
   char log[MLOG_LINES + 1][MLOG_LINE_LEN + 1];
-  uint8_t start_offset;
-  uint8_t lines;
+  uint16_t start_offset;
+  uint16_t lines;
   uint64_t last_update;
 } VitaChiakiMessageLog;
